Warn about location/dispersion mix-up in adonis
I was a bit reluctant to add this note, as this is a known issue:
Marti Anderson tells this in the paper where she introduces the method,
and adonis also seems to be much more robust than mrpp or anosim.
However, this completes the changes due to issue #3 in github.