Bias correction in betadisper with unequal sample sizes
HI Gavin, Jari, and Marti,

I'm writing to inquire about the importance of unequal sample sizes in
deriving and comparing beta diversity estimates across different groups.
My phd has focused on coral reef fish communities and in response to an
experiment with unequal sample sizes, a reviewer recently raised the
issue of unequal sample sizes in biasing beta diversity estimates when
comparing across groups. I've been working on this project with my
collaborator Ben Bolker (cc'd here). A numerical simulation he
constructed suggests that there is indeed bias in estimates of beta
diversity when sample sizes are unequal and sample sizes are small.

The solution that Ben converged on was a sqrt(n/(n-1)) correction factor
that can be applied to estimated distances to centroid for each group
(i.e. analogous to the correction commonly applied at the for sample
standard deviations).

 We attach a PDF file showing how we guessed this correction factor
(which makes sense in hindsight, although I would have difficulty coming
up with a rigorous derivation).

 We also attach a diff file from the current vegan version on r-forge
(r2127) which includes an option to correct the bias in this way (as
well as a one-line fix for an edge case, where there is only one
non-zero eigenvalue -- we came across this situation in the course of
running simulations, although it's probably not practically important
much of the time).  We think this should also work in the presence of NA
values, although we haven't tested it in that case.)