more canonical way of writing adipart.formula (also fixes match.call name)
Function adipart.formula() duplicated the full analysis of adipart.default.
The only difference was the interpretation of the input data. Now
adipart.formula() only interprets the formula and then calls
adipart.default() with the interpreted matrices.

This is also a maintenance issue: now things need be fixed only in
one file, and the methods do not randomly drift apart.

This also means that now the "call" from match.call() is always
set by adipart.default and will be generic adipart() instead of
adipart.formula().