oecosimu made more similar to adipart, multipart, hiersimu
oecosimu also returns "call" attribute
print.oecosimu displays "call" and can also display output of
adipart, multipart and hiersimu. However, print() of adipart and
multipart also shows some information on chosen options that is
not displayed by print.oecosimu.

The plan is to use print.oecosimu() for all similar functions
(adipart, multipart, hiersimu), and this is the first step to
that direction.