add labels.envfit()
The purpose is to allow setting names in plot.envfit(). The first
step is to add labels() function that shows the labels currently
in use. This is usually a list of two vector items, called "vectors"
and "factors", but now this is dropped to a vector if one of these
items is NULL.

This changes NAMESPACE and hence version number goes up.