Update scores.default.R
The modification allows one to work with univariate matrices (containing just one column) without the need to specify the choices argument. Now if x is numeric and contains just one column the scores.default function returns exactly the same x, without returns an error. This modification allows the use of the procrustes function with univariate matrices.