Remove a hack for 1-row scores given as a vector in scores.default
Solves issue #22. This reverts the hack of 2010-05-27 (R-Forge svn
commit r1196) that was supposed to make 1-row scores more robust.
It looks now that the hack would fail both when the 'choices' are
not given and when 'choices' are effective and only take some of
the columns. I cannot remember what triggered this hack, but now
it looks useless.