Stage 1 of fixing R CMD check warnings with current R
Removes tens of  warnings of "no visible definition" for
external packages, but still leaves this one warning:

* checking dependencies in R code ... WARNING
'library' or 'require' calls not declared from:
  ‘MASS’ ‘cluster’ ‘rgl’
See the information on DESCRIPTION files in the chapter ‘Creating R
packages’ of the ‘Writing R Extensions’ manual.

Wording is not clear, but I assume require(MASS) should *not*
be used (like it is done now), but we need MASS::isoMDS().