New internal function to get permutation matrix from variable input
In vegan, permutations can be defined either as a single number
giving the number of permutations (plus a strata argument defining
blocks), or as a how() structure defining the control in shuffleSet,
or they can be be given directly as a permutation matrix. This
function takes any of these alternatives and returns a permutation
matrix. The implementation is based on the current code in
permutest.cca.