SSarrhenius works when some observations have zero species
Selection of starting values failed when some observations had zero
species, because we initialize as lm on log(S). The actual fitting had
no problem with these same observations. We now use pmax(S, 1) to get
starting values -- which of course can be a worse now.

The problem was reported by Sara Landi of Siena.