Introduce generalized (parametric) dispersion weighting
This is a new unpublished method. Original method uses Pearson
Chi-square for a Poisson model with one factor and assesses the
significance using multinomial simulation ("c0_ind" nullmodel).
The new method estimates the same Pearson Chi-square for any glm
model and evaluates the significance parametrically. Otherwise the
models are identical except that the original method ignores
completely the factor levels where species in assessing the number
of degrees of freedom, but the generalized method defines the
degrees of freedom in a standard way (number of observations minus
number of model coefficients). As the degrees of freedom is used as
divisor for weights, this will influence the strength of downweighting.