remove u.eig, v.eig and wa.eig items from cca.default
The cca.object was documented in vegan 1.6-10 (Sep 26, 2005), and
in that document we warn that there is "no guarantee" that these
itmems will be kept in future release. In commit bddfea39 (Mar 5, 2013)
we use harsher tone and said that the "item will be removed". However,
that harsher text has not been in any release version so that this
may be an abrupt change.