esttimateR: use bias-corrected Chao1 only when doubletons = 0
Previously we always used bias-corrected form of Chao1, but
it seems that the normal biased form should normally be used, and
only when it cannot be estimated (because we divide by the number
of doubletons).

The sd of bias-corrected form is made NA in this commit: fixed in
the next one.