Use EstimateS formula (eq6) for the variance
The EstimateS web page eq6 formula is actually incorrect: it is
derived on the partial derivatives of classic Chao, but the terms
of the final result are replaced with the terms of bias-corrected
Chao. However, numerically the error is only about 1% of the correct
form and therefore we resent to use this widespread approximation.