Do not set strata if they appear to have been set earlier
anova.cca generated permutation matrix with strata in
getPermuteMatrix and passed that on to other functions (like
permutest.cca). This was broken in efd575f1 which passed
strata from anova.cca (unnecessarily), but we fix it in
getPermuteMatrix so that the solution is more general (and
strata are to be deprecated in the next releases in any case).