Better fix than 83d3bf1 and cca3f27 for resetting labels in envfit
Previous fix only looked at the null of internal labels, but these
can be null because the item (vectors or factors) was missing, or
the item was there but names were missing. Fix cca3f27 failed
in the first case and 83d3bf1 in latter case. Now we check for
the existence of factors/vectors, and if only one them is present,
we reset it labels whether they existed or were null.