More exact derivation of approximate variance of bias-reduced Chao
The common texts (and EstimateS web page) use approximate estimator
for bias-corrected Chao estimate. This uses derivation for the basic
estimator but replaces terms with those of bias corrected form
(a1^2 -> a1*(a1-1), 2*a2 -> 2*(a2+1)). The used form is unpublished
as such, but it follows exactly the derivation as given in Chiu et al.,
web appendix. The numerical differences between forms are tiny: in
BCI the maximum difference is ca. 0.02%, and in the 'mite' data up to
12% -- it seems that difference vanishes as the sample size (no. of
individuals) increases.