mvpart was archived on Dec 13, 2014: Rd cross-reference was removed
Brian Ripley sent an email on Dec 14, 2014, saying:

Its checks are now giving warnings:

* checking Rd cross-references ... WARNING
Unknown package ‘mvpart’ in Rd xrefs

That package is not mentioned in the DESCRIPTION file (it should have
been), and has finally been archived.

Please submit an update correcting this.

This was also raised as issue #90 github.