circumvent getSaveFile file extension bug in Tcl/Tk in Windows
Tcl/Tk documentation says that getSaveFile should automatically
append an appropriate file extension if none is given, but that
can be overridden with defaultextension. This happens in Linux
and Mac, but in Windows no extension is added to the plain name.
However, if defaultextension has any value in Windows, the correct
extension is added to the plain name instead of using the value
of defaultextension.

This is clearly a Tcl/Tk bug, and has been reported several times
on the Web. If Tcl/Tk in Windows is fixed, this kluge should be
removed.

The fix should be  tested in Windows.

(cherry picked from commit 8b28eaf19884922e96a0bd4c7620b5a788263d42)