R CMD check --as-cran checks now urls, and two failing ones found
In cca.Rd we referred to Dave Roberts's old USU pages, and the new
pages no longer have the code we refer to. In sipoo.Rd the original
Atmar & Paterson web page has evaporated, but the actual data
were since dug out form the literature reference and we use it now.