Protect users against themselves: some give dissimilarities to rda/cca
Input of data as dissimilarities was silently accepted in rda
and cca. The dissimilarities were used as a regular community
matrix, but the results are non-sense. I have recently seen one
example of this misuse, and hints to this in another query.

(cherry picked from commit 4731b22acbad6c9212c95cba8f3076fdeb21e5ee)