More robust scoping in capscale
capscale updated formula with a new data calculated within the
function, and the environmental of the formula vanished after
exiting. This could cause problems when the result was reused in
anova.cca etc, or when capscale was embedded in other functions.

tests/vegan-tests.R failed in all capscale cases before this fix.