Partial component is only added to the explained part of constraints
Earlier we added the variation explained by the partial component
both to the constraints and to the residual component. For a partial
model 'm' the following gave values > 1 because partial was used twice

goodness(m, mod="CCA", sum=TRUE) + goodness(m, mod="CA", sum=TRUE)

but now this gives 1 for all species (all for all sites).

The problem was reported in R-sig-ecology by Christoph Eberhard in
https://stat.ethz.ch/pipermail/r-sig-ecology/2015-March/004936.html