predict.specaccum uses analytic models with 'newdata'
Earlier we interpolated from values fitted to integer number of
sampling units, but now we can calculate the expected species
richness for fractional number of sites with "exact", "rarefaction"
and "coleman" methods. For "exact" and "rarefaction" we now use
equations based on gamma() functions instead of choose() because the
former can handle non-integer data.