fitspecaccum for weighted specaccum consistent with unweighted
fitspecaccum() scales weights to averaged sites so that max
accumulated weight == number of sites. In this way predict takes
newdata as sites and specslope gives the correct slopes. Earlier
we had to give the effort and multiply the slope by the average
increase in effort per site.

plot.fitspecaccum and lines.fitspecaccum were changed accordingly.