Really deprecate commsim
commsim was deprecated in release notes of vegan 2.2-0, but
this triggered warnings in dependent packages. To give time to
adapt to changes, no warning was issued. Now we issue a warning,
but it seems that other CRAN packages already chanaged their
usage and no problems were found with tools::check_packages_in_dir()