ellipse on points on line has zero area
Eigenvalues of covariance matrix are positive and if the points
are on line, the second eigenvalue is zero. In theory. In practice
the zero eigenvalue can be numerically estimated as a tiny negative
number, and then area will be NaN. Previously this was fixed for
two points (which are necessarily on line), but this will take care
of three or more points on line as well.