New dbRDA function by Pierre Legendre
<Pierre.Legendre@umontreal.ca> on  Fri, 12 Jun 2015 13:18:40 +0200:

"I wrote a small function called dbRDA.D a few months ago, and used it
in two papers at this point. This function computes the dbRDA F-test
of significance between response data represented by a Euclidean or
non-Euclidean ***dissimilarity matrix*** and a matrix of explanatory
variables, using the method of McArdle and Anderson (2001).

For a Euclidean response distance matrix, this is the exact equivalent
of computing PCoA on the distance matrix, then computing RDA and its
F-test of significance. The calculation is completed using directly the
D matrix without having to do the PCoA.

The beauty of this method is that it also works for non-Euclidean distance
matrices as the response data.

In many cases, all we want is the p-value, R-square and adjusted R-square.
This is where this method is most useful. The function can also (optionally)
produce the RDA eigenvalues and ordination axes."