zap.ind is dangerous in printCoefmat: do not use it in print.cca
'zap.ind' argument first rounds numbers to the given 'digits' and
then pads the numbers to the same numbers of decimals by adding
trailing zeros. This is misleading and wrong.

'zap.ind' is also used in print.permutest.betadisper and
print.vectorfit, but there the numbers may be in the same scale
originally and therefore wrong padding is rare (but needs
checking). In print.cca proportions and inertia are ofen in
different scales, and eigenvalues get wrong trailing zeros.