Circumvent shuffleSet(2) bug in CRAN release of permute  0.8-5
shuffleSet(2) should return a 1x2 matrix (row), but it returns
a 2x1 matrix (column). This is fixed in GitHub for permute, but
not yet released to CRAN. Because of this, commit 1befc45e6
broke nesteddisc.

There are some other permute:::shuffleSet kluges that should
be removed after new permute becomes available in CRAN.