Do not use unnecessary (w)cmdscale but just GowerDblcen dissimilarities
dbrda used (w)cmdscale only to get Gower double-centred matrix
and all eigenanalysis was unnecessary burden. Now we only transform
dissimilarities. With this change, we also drop the add argument that
was able to euclidify the matrix. This option is still available in
capscale(), and its results are identical to dbrda() when there are
no negative eigenvalues (like after add=TRUE): use capscale() if
you want to use the option and you get the same results.