changing scaling to numeric could fail if scaling already was numeric
In most functions setting scaling in cca/rda/capscale the scalingType
was wrapped within if(is.character(scaling)), but when this was
omitted, the no-op of retaining numeric scaling failed. This crashed
summary.cca and tolerance.cca where the wrapping was not done. Now
the test was moved within scalingType() which returns given numeric
scaling unchanged. This also means that if-wrapping could be
removed.

This was detected in tools::check_packages_in_dir for BiodiversityR.