more forgiving test for tree reversals
We used is.unsorted() to see that the merge heights were non-decreasing
and the tree has no reversals. With tied merge heights, it could happen
that the difference of tow tied values is a tiny negative number (-3e-15)
and the function assumed a reversal. Now we need a decrease of more than
sqrt(.Machine$double.eps).

The fix was made only for functions that signal error with tree
reversals; treeheight still uses is.unsorted() but there the only cost
of wrong diagnosis is to use a slower method to measure the tree height.

(cherry picked from commit 215fedafe4a6c2d5548b91e3e91d945257695993)