all specpool indices "work" when there is only one sampling unit
No extrapolated richness can be estimated when there is only one
sampling unit, but now now specpool() returns the observed number
of species and zero standard error for all such cases. It used to
fail with Chao index and give NaN for the SE of jack2.

This was reported in StackOverflow in
http://stackoverflow.com/questions/34027496/error-message-when-using-specpool-in-vegan-package