Add argument 'try' for minimum number of tries in metaMDS
Sometimes a repeated local optimum is reached in metaMDS, but
with 'try' arg for the minimum number of tries, the search still
continues, and a better (global?) optimum may be found.

Currntely this old 'trymax=20' is kept, and the new argument defaults
to 'try=6'. These limits may be increased. However, even with this,
some examples() will get beyond a local optimum in metaMDS(dune) and
a better solution can be found. We'll consider having still higher
default limits for 'try' and 'trymax'.