add option of thinning to quasiswap
quasiswap is non-sequential method, but we added thinning to avoid
bias. quasiswap does either ordinary trial swap, but some-of-squares
reducing swap transforms (2,0,0,1) -> (1,1,1,0) and (2,0,0,2) ->
(1,1,1,1) or matrices of fill=3 or 4, and these matrices may be
over-represented in the result. In particular, the analysis ends
with this ss-reducing step. With thin > 1 we can do several trial
swaps after ss-swaps and this will reduce (or remove) the bias.