Add method to compute adjusted R2 for cca
This method uses the permutation approach suggested by Peres-Neto et al 2006. It does currently have limitations that need to be fixed. Namely the function requires that the input 'x' be parameterized by the function cca using a formula rather than individual matrices, and this code assumes the cca argument 'data' is specified.