allow user interrupts in compiled nullmodel code
several C functions for null models had a potentially infinite
while() or for(;;) loop with exit only when the condition for
final solution was meet. It was possible to construct matrices where
such a solution did not exist, and the null model simulation went into
infinite loop. This commit allows user interrupt in interactive
sessions.

This is the alpha commit: handling of user interrupts must be made
less disturbing.