argument FUN rarely worked in plot.spantree
The code used FUN(d, y) where y is the initial configuration, and
this worked only for methods that had this second argument. This was
pretty much an empty set. Now we use this form only when FUN is sammon,
and for other methods try FUN(d).

This looks pretty much as a useless argument, since the initial
configuration can be given as argument ord, and there is no need
to define a function to calculate that ord internally.