more robust 'alternative' in summary.permustats()
Internal 'alt' variable (from 'alternative' argument) could be either
numeric or character string, but then switch() using 'alt' should use
labels in the correct numeric order. This was not done earlier, and was
fixed in 351c5b812. Now we also make 'alt' always to character so that
there is no danger of mixing order of labels.