try to save the solution so far if the user interrupts metaMDS
Do not abort immediately, but return the current solution if
user interrupts (ctrl-c) metaMDS during MDS iteration. The rescue
is not guaranteed: if ctrl-c'ed at other stages than during NMDS
iteration, abort is immediate and nothing will be returned.

This commit implements the simple test case for non-parallel
runs: with parallel runs abort is immediate and all will be lost.