update examples to use improved summary in R-devel (after r71150)
R used to round (hard) results in summary and this gave misleading
trailing zeros if decimal point was not in the same position. The
handling was improved in R, and now we update the vegan example to
use that improved handling.

The results match now R-devel: the release version still has the
confusing rounding.

The svn entry for this change is
------------------------------------------------------------------------
r71150 | maechler | 2016-08-25 22:57:19 +0300 (Thu, 25 Aug 2016) | 1 line

summary.default() no longer rounds by default; just *prints* rounded