Faster minimum terms in designdist
Minimum terms are found using .Call interface to compiled C code,
and run now 15% to 40% faster in microbenchmark. The main reason
for the speed-up is that C does not check and skip missing values
but gives NA for any dissimilarity involving NA data. This is
consistent with "quadratic" and "binary" terms. The R interface is
also less hackish and this gives some extra speed-up.