remove dimension checking from random number generation
millions of random numbers may be generated with one matrix (e.g.,
in quasiswap) with same dimensions, and it is wasteful to repeat
check for dimensions every time when the dimensions do not change.
This should be checked before entering the loop or in R before
calling C (like it is done now in make.commsim after 96fbde080 and
github issue #177).