test a potentially faster alternative for trialswap
Use one test to get the two out of sixteen 2x2 matrices that can
be swaped. Tests indicate that this indeed is sligthly faster or
as fast as the old code. We had a test for those 6 of 16 matrices
that only had two ones and include those 2 that can be swapped, and
for sparse matrices (= low fill) this seemed to do the same job, but
for denser matrices we are no somewhat faster.

This commit has a non-.Call interface both the old code (nullmodel
"tswap") and new code (nullmodel "ftswap") for testing purpose.