change trialswap to use the new snappier code
the new code was called as model "ftwsap" and old as "tswap" from
commits a9a3876db to d5b198715 which can be used for testing and
comparing methods. New code is usually slightly faster, but in sparse
data the old test also skipped most of the analysis and difference
is slight. There may be a larger difference by switching form static
inline function to using a macro in generating two random numbers.

Old and new code give identical results with the same random seed.