Revert "use similar bitwise shift and OR in swap as previously in trialswap"
This reverts commit 495cae4c69fcb9b38d157e1ec6e4cda750f6f4f9.

Old code is more transparent, and it is probably as fast, and
with sparse data it may be faster: the first test skips the rest
of the loop unless there are two occurrences in 2x2 submatrix,
and in sparse data this is the most common situation. In dense data
the test is true at max 37.5% cases which means that the number of
tests will be about equal to the current alternative with bitwise
operators.