add function qrX to reconstruct X from its QR decomposition
Also added SEXP function to test this from an R session:
.Call("test_qrX", sol$CCA$QR) should give the same result as
qr.X(sol$CCA$QR) except for pivoting.

This is needed in weighted ordination (CCA), and can be called from
the C function. In current parallel processing only one permutation
is handled and this puts an overhead to this function. Either we
define data before calling the getF function or design a better
parallel processing for larger numbers of permutation in time.