add function to test QR decomposition from C
The R base::qr() does not actually use the original Linpack
function, but an edited version. From the user-point of view the
main difference is that the Linpack version does not return rank
and it does pivoting differently. The test function can be used
to see if we can still apply Linpack. The test function can be
called as .Call("do_QR", x, 1), where x is the (centred) model
matrix, and the second argument controls pivoting (1: do pivoting,
0: keep the original order of columns). The do_QR function returns
an object of class "qr" which technically can be used by R qr.*
functions, but rank may be wrong (it is the number of columns instead
of the real rank), and the pivoting is done differently.