begin refactoring of constrained ordination methods
constrained ordination methods (cca, rda, capscale, rda) share
much in common and could be implemented with shared code with some ifs
for different methods. The major difference is in the setup of the
dependent (community) data. This commit implements first versions
of setting up community data and provides a driver function ordConstrained
that finally should be embedded in constraining ordination functions, but
is now exported for testers.