revert reweighting of constraints in CCA in permutest.cca
Since vegan release 1.6-10 (Sep 26, 2005) we have re-weighted
constraints and refitted QR after each permutation in CCA.
This was a conscious decision that broke Canoco 3 compatibility.
Now this decision is turned over, and the results are compatible
with old Canoco.

The change was made here in R code: C version will probably just
fail.