re-instate old scaling of internal results in RDA/PCA for support functions
The new ordConstraint scales data so that the eigenvalues directly
decompose the variance, whereas earlier we analysed non-scaled data
and only scaled the eigenvalues. Several support functions should
be changed to handle this, and therefore we now make rda.default
to back-scale internal data (Xbar, colsum) to the original data.
Eventually we should fix the support functions.