faster colsum term for rda
rda used to be slower than cca, and this was mainly due to its
slower ordHead function which used sd to get the standard deviations
of variables. The internal data were already centred and therefore
we could switch to 1000x faster colSums, and now rda is faster than
cca (like was expected).