inverse policy to define 'adjust' in dbrda & capscale
we used to define 'adjust' as the term that is needed to cancel
variance approach in rda to get the sum of squares for normal
dissimilarity indices with ceiling at 1. If we had such an index,
adjust was sqrt(n-1), but for Euclidean distance and alike we had
adjust 1. Now adjust is the divisor used for the adjustment, so that
it is sqrt(n-1) for variances and 1 for sum of squares. This is done
in anticipation of using ordConstrained with specific init functions
for capscale and dbrda.

Support functions are not yet adapted to this.