downgrade capscale: no longer parallel distance-based components
From vegan release 2.4-0 (and commit 150bfa6 in development)
capscale has added dbrda-like decomposition of original dissimilarities
after finding the ordination solution. The ordination solution was
based only on real dimensions (with positive eigenvalues), but the
dissimilarity components also included negative eigenvalues. In this
way the permutation tests etc were similar in capscale and dbrda
although the analyses differred. Now we go back to the pre-2.4-0
behaviour of skipping the parallel analysis and only reporting the
decomposition of real components.

The change is discussed in github issue #217.