anova and permutest failed for betadisper with bias.adjust = TRUE
With bias.adjust=TRUE we used table() to get the counts for groups,
and the result had class "table" with special as.data.frame() method.
We used data.frame as the internal structure to pass data to anova
and permutest, and this was distorted with as.data.frame.table().

Fixes issue #219 in github.