inertcomp gains option to scale entries to sum to unity
This will give "local contributions to beta-diversity" (LCBD) and
"species contribution to beta-diversity" (SCBD) of Legendere &
De Caceres (Ecology Letters 16, 951-963; 2012) with similar scaling
when the dissimilarity is appropriate in dbrda or data were properly
standardized in rda. The default gives contributions that sum up
to beta-diversity, but scaling to unity is used in the numerical
code by @legendre.