biplot scores need reversed scaling to other scores
Normally we multiply with the eigenvalue scale, but biplot scores
need be divided. Originally the scaling was introduced in commit
099bf7b050d after being prompted by Daniel Borcard and
Pierre Legendre (@legendre) and this commit will change that
behaviour. The scaling of this commit makes biplot arrows consistent
to fitted vectors (envfit) and ordisurf(..., knots=1) when applied
to lc scores with the same scaling.