as.mlm does not force regression through the origin
In as.mlm() the intercept will be ~ 0 because both the response
and the predictors are centred. Numerically estimated intercept
will typically be a tiny non-zero number which causes ugly print
of the coefficients. Forcing regression through zero gives the
same fit with more beatifully printed coefficients, but it gives
wrong residual degrees of freedom and therefore in summary() the
standard errors of coefficients are too low, t-values too high,
P-values too signficant and adjusted R2 too high. However, the
whole summary statistics are misleading because they are based on
the variance of WA scores, and have no meaningful interpretation
for the data.