take length of weights from the length of colsum in dbrda/capscale/rda
The colsum can be NA and of length 1 when dbrda/capscale was called
with dissimilarity matrix instead of raw data + dist=argument, but
in that case the real dimensionality of columns are also unknown and
old code gave wrong length anyway. The colsum is always filled when
the number of columns is known.