fitted.capscale uses ordiYbar and internal structures
type = "working" was earlier found from eigenvectors, but now we
use internal Ybar for pCCA, CCA and CA. The column dimensions
change: now they are equal to the number of real dimensions like
in the input data, but earlier they were equal to the number of
positive eigenvalues in the component. The resulting distances are
the same. Now we also have fitted for pCCA component.