take care that numerically approximate 1 is exactly one
We divide with 1-1/N2, and at the border case we can have N2=1
(when species only occurs in one SU, or SU has only one species).
N2 is a floating point number and it can be slightly inaccurate
1, and if it is 1+EPS we end up dividing with EPS. This can give
results which look acceptable but are wrong -- the result should
be 0. The previous version used zapsmall(N2) which avoids this
problem, but it zapsmall() rounds all N2 values (not only approximate
1's) and since we divide, rounding adds error.