document defunct functions forever
We used to document .Defunct functions in man/vegan-defunct.Rd as
long as the .Defunct() function entry was still in vegan, but
remove the documentation when all traces of the function were removed.
The model in stats-defunct.Rd and base-defunct.Rd says that the
documentation entry should be kept in the Rd file even after the
actual function was removed.