do not set non-standard attributes on dist object as default
It is considered dangerous to set upper = TRUE, diag = TRUE on
default return object, because this makes users believe the output
is "matrix" instead of "dist". If desired, the attributes can be
set in call as "..." are passed to distfun which will set these
attributes in the reuslt  (if working in standard way like dist and
vegdist).