make tol dependent on the first eigenvalue
The condition number (kappa) is the ratio of largest and smallest
kept eigenvalue, and using this as a criterion guarantees that
kappa of the kept eigenvalues will not exceed tol. Even after this,
the Mahalanobis transformation usually fails when the number of
rows is lower than the number of columns.

(cherry picked from commit c3c28e51a66de9ac50811add5def906e97dceb1e)