make tol dependent on the first eigenvalue
The condition number (kappa) is the ratio of largest and smallest
kept eigenvalue, and using this as a criterion guarantees that
kappa of the kept eigenvalues will not exceed tol. Even after this,
the Mahalanobis transformation usually fails when the number of
rows is lower than the number of columns.