remove more items in backtracking if there was no improvement
Comparison against R code of backtracking showed that the C code
produces stronger bias. Two main differences are that initial filling
in R code is proportional to probability from marginal totals and equal
in C, and progressively more picked items are dropped in backtracking
if no success was made in R, but only one item was dropped in C.
It seems that the latter is more important: removing only one item
is not sufficient to break blocks. This commit implements similar
dropping scheme as the R code and removes progressively more
items (up to 4) if no success was made, and resets to drop one after
each improvement. This commit does not (yet) implement restoring
previous solution when the solution gets worse. (Debug infromation
on current number of picked items is printed.)