do not go backwards in backtracking but reset to previous solution
If the 'fill' decreases from previous ones, reset the solution to
the previous one with achieved fill. This is a simple implementation
doing a lot of memcpy in and out at every step. My intention is
to work out parallel implementation with index swapping to reset
the solution (and doing nothing in other cases) in the parent
branch.