add static function for boosted quasiswap: a variant of curveball
Function works by picking two random rows, and finding all species
that individually can be quasiswapped, and for both rows quasiswap
equal number of species up and down. During this it also makes a
number of sum of squares reducing quasiswaps and continues as long
as there are items > 1. Normal quasiswap inspects 2x2 submatrices,
and finding the last > 1 item in swappable combination takes a lot
of time, and working with two rows is potentially faster.

The function is now syntactically correct, but there is no user
interface yet, and I don't know if it also works correctly...