study a very greedy method of quasiswap: always pick >1 item
This is even greedier alternative than dd15ed5ac: we always pick
>1 cell as the corner of the 2x2 matrix. The function is a minor
variant of quasiswap. Instead of assessing sum-of-squares, we have
a vector of cells indices > 1, and we iterate till the vector is
empty. There is much less bookkeeping than in dd15ed5ac: after
each step we only inspect if any decremented item becomes 1 and
will be removed from >1 vector, or any incremented items becomes 2
and will be added to the >1 vector (this can happen).