add thinning to greedy quasiswap
Each successfull greedy step will produce a 2x2 matrix with fill
3 or 4, and the result matrix is heavily biased. With thin, only
every thin-th step is greedy, and we have ordinary quasiswap steps
to remove the bias. In tests even modest thinning corrects the
bias in checkerboards: thin=10 was often sufficient, meaning that
10% of steps are greedy.

Without thinning, the method is much faster than boosted quasiswap,
but with sufficient modest thinning it is about as fast.