do not use too many random numbers in trialswap: they are slow
Getting random numbers takes most of the time in trialswap: we
use four in each cycle (two rows, two columns), but most were
wasted because we could not swap the submatrix. This commit reduces
the random numbers to three in worst cases and often to two: we first
find the matrix index a of m[a] first, and then find a different column
for index [b]. If m[a] == m[b], the matrix is unswappable we bail out
and start a next cycle. If m[a] != m[b] (i.e, we have 0 and 1), we
find the second row and items m[c] and m[d].

This sped up 1e9 cycles (100000 simulations with thin 10000) from 52 sec
to 29 sec in my Ubuntu desktop and from 68 sec to 48 sec in my Macbook
Air.