get first two rows in one column to speed up calculations (sligthly)
We use at maximum 3 random numbers to find the 2x2 submatrix, but
we use only two if the first two items give a matrix that cannot be
quasiswapped (both items 0) or swapped (both items 0 or both 1). The
average number of used random numbers is mean(3 - (1-f)^2) for
quasiswap and mean(3 - (1-f)^2 - f^2) for swap, where f is the
relative col/row fill. We used to find the first two items in the
same row (sampling unit), but if we find them in the same column
(species) we need fewer random numbers, because species fills
(frequencies) are usually more variable than row fills (alpha
diversities):

Table: average number of random numbers for one cycle

                       Same row /earlier      Same column / now
          total fill   qswap     swap         qswap    swap
sipoo     0.233        2.37      2.29         2.35     2.24
BCI       0.403        2.64      2.48         2.54     2.27
bryceveg  0.080        2.15      2.15         2.14     2.13

PS. We can have a scheme that always uses only 2 random numbers
per cycle, but I'll test it later. This gives a small speed-up.