use get2x2 in quasiswap
First test with BCI indicate that this is *slower* than the current
scheme. This is surprising, as profiling of compiled code suggests
that 2x2 scheme is faster for BCI (though not for bryceveg).
Even if the code is slower, it is much more readable as it tries
not be so clever. However, this may be reverted if tests show
consistently that the old (3minus) code is faster.