handle zero-rank conditions in partial ordination
If conditions explain nothing in partial models, and their tot.chi
is practically zero, it is set to 0 and rank is also set to 0 (as
these are zero-rank models). The behaviour changed from previous
vegans, where the rank of conditions was still given although the
partial component was actually of zero rank. Several other functions
need the rank of conditions, and it must now be found from
pCCA$QR$rank instead of previous pCCA$rank when pCCA$rank==0.