rda(Y, X) failed if X was a single variable and dimensions dropped
This happens only when rda() was called w/o formula. It cropped out
wit R CMD check analogue_0.17_0 with vegan_2.5-0 and running examples
failed in compare(). The ultimate error was in residLen. Now rda.default()
takes care that input is a matrix, similarly as in cca(). It would make
sense to move both cca() and rda() tests to ordConstrain(), but not yet.